+++ R/get_private.R
@@ -1,0 +1,4 @@
+
+get_private = function(x) {
+  x[[".__enclos_env__"]][["private"]]
+}