+++ tests/testthat/test_map.R
@@ -204,3 +204,21 @@ test_that("detect", {
 test_that("pmap does not segfault (#56)", {
   expect_error(pmap(1:4, function(x) x), "list")
 })
+
+test_that("walk", {
+  .x = list(a = 1, b = 2)
+  sq = function(x) x^2
+  expect_equal(walk(.x, sq), .x)
+})
+
+test_that("iwalk", {
+  .x = list(a = 1, b = 2)
+  sq = function(x, name) x^2
+  expect_equal(iwalk(.x, sq), .x)
+})
+
+test_that("pwalk", {
+  .x = list(a = 1:3, b = 2)
+  f = function(a, b) a + b
+  expect_equal(pwalk(.x, f), .x)
+})